+++ packages/contentful--create-contentful-app/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.6](https://github.com/contentful/create-contentful-app/compare/@contentful/create-contentful-app@0.15.5...@contentful/create-contentful-app@0.15.6) (2022-03-09)
+
+
+### Bug Fixes
+
+* **deps:** bump inquirer from 8.2.0 to 8.2.1 ([07167a6](https://github.com/contentful/create-contentful-app/commit/07167a61ad412bb6d76ab86c52984c865570d681))
+
+
+
+
+
 ## [0.15.5](https://github.com/contentful/create-contentful-app/compare/@contentful/create-contentful-app@0.15.4...@contentful/create-contentful-app@0.15.5) (2022-03-02)