+++ packages/contentful--react-apps-toolkit/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.4.2...@contentful/react-apps-toolkit@0.5.0) (2022-03-10)
+
+
+### Features
+
+* useFieldValue [EXT-3492] ([#923](https://github.com/contentful/create-contentful-app/issues/923)) ([3f1c25f](https://github.com/contentful/create-contentful-app/commit/3f1c25fca2b6458e2bd393efccdcedcb679f5754))
+
+
+
+
+
+## [0.4.2](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.4.1...@contentful/react-apps-toolkit@0.4.2) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/react-apps-toolkit
+
+
+
+
+
+## [0.4.1](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.4.0...@contentful/react-apps-toolkit@0.4.1) (2022-03-10)
+
+**Note:** Version bump only for package @contentful/react-apps-toolkit
+
+
+
+
+
+# [0.4.0](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.3.0...@contentful/react-apps-toolkit@0.4.0) (2022-03-04)
+
+
+### Features
+
+* [EXT-3491] add useCMAClient hook ([#917](https://github.com/contentful/create-contentful-app/issues/917)) ([2c53ddf](https://github.com/contentful/create-contentful-app/commit/2c53ddf2a2ffd341fadd8aa59425b8f1b07b6d7b))
+
+
+
+
+
+# [0.3.0](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.2.1...@contentful/react-apps-toolkit@0.3.0) (2022-03-04)
+
+
+### Features
+
+* [] add timeout and console warn ([#919](https://github.com/contentful/create-contentful-app/issues/919)) ([9afcad6](https://github.com/contentful/create-contentful-app/commit/9afcad6f26ae7f0023942a8fa2110096f6cb3fd5))
+
+
+
+
+
 ## [0.2.1](https://github.com/contentful/create-contentful-app/compare/@contentful/react-apps-toolkit@0.2.0...@contentful/react-apps-toolkit@0.2.1) (2022-03-03)